@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
+    "@mui/icons-material": "^6.4.11",
+    "@mui/material": "^6.4.11",
     "@zxing/browser": "^0.1.5",
     "@zxing/library": "^0.21.3",
     "qrcode": "^1.5.0",
-    "ra-i18n-polyglot": "^4.16.20",
-    "ra-language-finnish": "^4.13.0",
-    "react": "^17.0.2",
-    "react-admin": "^4.16.20",
-    "react-dom": "^17.0.2",
+    "ra-i18n-polyglot": "^5.7.3",
+    "ra-language-finnish": "^5.7.0",
+    "react": "^18.3.1",
+    "react-admin": "^5.7.3",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.30.0",
     "styled-components": "^4.4.1",
     "uifx": "^2.0.7"
