@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Button, Title, useNotify, GET_LIST } from 'react-admin';
 import { Navigate } from 'react-router-dom';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import TextField from '@material-ui/core/TextField';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import TextField from '@mui/material/TextField';
 import { juniorProvider } from '../providers/juniorProvider';
 import { httpClient } from '../httpClients';
 import { STATE } from '../state';
