@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { Button, Title, useNotify } from 'react-admin';
 import { Navigate } from 'react-router-dom';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
 import { httpClientWithResponse } from '../httpClients';
 import api from '../api';
 import { STATE } from '../state';
